Changed Poll to use timeLimit as String and added pollClosingTime, also as a String.
ZonedDateTime was a bit tricky to get working with JSON,
so it is now being stored as a String instead.
Provide the String to ZoneDateTime.parse(), to use it as a ZonedDateTime object.

Added a new field variable, pollClosingDate, to hold the
date and time for when the poll should close after being activated.
Is generated based on current system time + the given time limit.
Should be set automatically and appears to return the value when queried,
but might return null in JSON when submitted,
will investigate more if it becomes a problem.

Fixed the setters, getters and toString() to represent the two variables.