Now using IDs of Polls instead of Poll objects.
Using Lists of Polls, created an infinite loop
in the JSON response from requesting a poll.
The JSON returned the poll, with the creator
containing the poll itself in its
list of created polls. Resulting in the poll and
creator repeating each other endlessly...

Changed pollsVotedOn and pollsCreated
from List<Polls> to List<Long>.

Renamed the variables and methods
to reflect the changes.