Security (#8)
* Added variable canEdit in Poll,

Used to determine if poll is still editable, and not yet published.

* Added file for testing TimeDuration.java

Made test for checking that converting a String to TimeDuration has the time units specified in the String.

* Added tests for validating time units

Added tests for checking if a TimeDuration has the correct time units.

* Added more tests

Now tests that a TimeDuration is created with the intended values.

Added some variables for making a standard String of time units, simplifying some tests.

Added some simple tests for getters.

* Added tests for setters, reducing and increasing.

* Added class for validating passwords

* Now uses autogenerated field for ID instead of username

Added getter and setter for new id field.
Added id to User.toString().

Using autogenerated values for id instead of username, help increase security by limiting places where sensitive data will be handled.

* usernames should be unique

* User credentials are now validated for new users

Added a logger for logging purposes
(does not save anything to file, for the moment)

Made a method for evaluating user credentials.
* Checks that the username is not in use
* Checks that the submitted password follows
  certain rules.
* Checks that the repeated password
  matches the original password.

* Now checks if password is among the most used

Added a file with 10 000 most commonly used passwords.
Added a rule to PasswordValidation, to check that
the password is not one of those 10000.

* Class for generating new passwords

Has a method for generating a new password,
and a method for replacing all existing users
passwords with newly generated ones.

Useful if stored passwords are compromised
at some point.

* Added loggers to all requests in Controller.java

* Repositories now injected via a constructor

Repositories now use constructor dependency injection,
instead of field dependency injection.

* Limits access to webpages and requests to authorized users.

Currently not very secure and should
be updated with the proper webpages
when they are ready.
Should also update request limitations
when final URLs have been decided.

* Keeps track of the webpages to manage

Should be updated when the real webpages are ready.

* Added some html files to serve as webpages for testing

Should be replaced by actual webpages
when they are ready.

* Added dependencies for security and testing purposes

Thymeleaf: for testing that the webpages work, might be removed at a later date.

Passay: Library for password generation and validation.

Lombok: for some useful annotations.

Spring Security and testing: Password encoders
and access limitations, and other security stuff.

jetbrains: for some useful annotations.

* Now accepts negative time durations from strings.

* ToString now has each time unit starting in uppercase

* Added many other things for git to ignore

Removed some duplicates

* Added more things for git to ignore

* Class for generating new passwords, using custom rules

Now logs usage information.
Added new custom rule for generated password.
Refactored the class.

* Added logging messages and a method estimating entropy

Method estimates entropy based on the rules
defining valid passwords.

Estimated entropy could be used to provide
users with feedback on passwords.

* Changed from testing.

* Constructor now updates pollClosingDate automatically

If time limit is not specified or it is infinite,
pollClosingDate will also reflect that
infinite deadline.

* timeLimit is now infinite, if not specified in setter

* Now permits all CRUD operations on all paths.

Not very secure, but works for testing.

* Now setting pollClosingDate automatically too!

Hopefully it works...
Time will tell if changes are required.

* Saved the derby log.

Should be removed from git in the future.

* Now using IDs of Polls instead of Poll objects.

Using Lists of Polls, created an infinite loop
in the JSON response from requesting a poll.
The JSON returned the poll, with the creator
containing the poll itself in its
list of created polls. Resulting in the poll and
creator repeating each other endlessly...

Changed pollsVotedOn and pollsCreated
from List<Polls> to List<Long>.

Renamed the variables and methods
to reflect the changes.

* Created a nested private class AdminTools in Controller

Can be used to perform administrative tasks
from a simple to use interface.

Right now only capable of replacing every users,
or a specific users, password with a
randomly generated one.

Might not be needed. Will see if
it should be removed in the future.

* Creating a new poll now requires a valid creator.

Creating a new Poll now checks if the username
of the creator exist in the database.

* Removed debugging messages

* Now hashes the user password.

* Added a dependency required by hashing algorithm

BouncyCastle: Needed for password hashing algorithm.

* Now measures the password hashing time and logs it

* Saved derby log again...

Co-authored-by: Tore Berven <tore.berven@student.uib.no>