Enabled HTTPS
* Now uses SSL with TLS to secure requests.

* Redirects all HTTP connections on port 80/8080
to 8443.

Requests in postman will need to be changed from
http://localhost:8080/
to
https://localhost:8443/
or else all requests will be disallowed or converted
to GET requests.
May need to disable SSL verification in postman
if it complains.

Some web browsers might complain about insecure
connections, due to invalid certificate.
To fix for local client, export a certificate from
the key, and import the certificate from
keystore to Java SDK path/*/cacerts
, which contains all valid certificates.