Now simplified and better than ever
* Can now choose to subscribe for vote updates,
or time remaining updates. Default is true for both.

* Removed main channel, as it no longer made sense
to include it.

* Removed sendMessage(), as it was kinda pointless
and confusing.

* As a consequence of removing sendMessage(),
the messages sent to the server have been simplified
to only include necessary JSON values for each
particular request.

* The Controller.java have also been refactored
to reflect the websocket changes.

* Added a handy message to websocketConnections.html
informing users to utilize the developer console for
interacting with the websocket.